Create Visualization of Traffic Vs Restaurant Counts
As can be seen in the scatter plot, there is no clear linear relationship between the number of restaurants within the specified radius and traffic flows in this limited dataset. The purpose of the exercise was to brush up on preparing and analyzing data for future use with machine learning.